add prepared stament helpers, - add table to oid vector helper - implement setup_prepared_statement correctly

+++ src/types.hpp
@@ -11,7 +11,7 @@ using ImdbID = std::string;
 
 struct TitleRating {
   ImdbID tconst;
-  double averageRating;
+  float averageRating;
   std::int64_t numVotes;
 
   POSTGRES_CXX_TABLE("title.ratings", tconst, averageRating, numVotes);
